:art: FORMAT (exhibition): move role and section into components
