+++ src/components/ExhibitionSection/index.tsx
