+++ src/pages/ForgotPassword/index.tsx
